stop using errno in die
As in the previous commit, it is probably not be a good idea to read
from errno here because it might have changed in the meantime and also
because Rust already gives you a way to pass error information.

Using the Display trait here will allow you to pass pretty much all errors
from the standard library straight through. This is useful when using
the Result::Err values that you get.

Also as per the documentation page for the ToString trait (which is the
same idea as Into<String>) says to prefer implementing the Display trait.